backend/requirements.txt
fastapi
uvicorn
google-cloud-tasks
google-cloud-firestore